Geodjango/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-import-export==2.4.0
django-leaflet==0.27.1
django-model-utils==4.1.1
django-pwa==1.0.10
django-rest-swagger==2.2.0
djangorestframework==3.12.2
docker==4.3.1
et-xmlfile==1.0.1
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
optional-django==0.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
tablib==3.0.0
tqdm==4.54.1
uritemplate==3.0.1
urllib3==1.25.11
websocket-client==0.57.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
uWSGI==2.0.19.1